refactor: move dom helpers to dom module
