+++ src/store/columnCoords.ts
@@ -114,29 +114,24 @@ interface ColumnCoordsOptions {
 
 export function create({ column, dimension }: ColumnCoordsOptions): ColumnCoords {
   return reactive<ColumnCoords>({
-    get contentsWidth(this: ColumnCoords) {
-      const columnLen = column.visibleColumns.R.length + column.visibleColumns.L.length;
-      const totalBorderWidth = (columnLen + 1) * dimension.cellBorderWidth;
-      const scrollYWidth = dimension.scrollY ? dimension.scrollbarWidth : 0;
-
-      return dimension.width - scrollYWidth - totalBorderWidth - dimension.frozenBorderWidth;
-    },
     get widths(this: ColumnCoords) {
       const { visibleColumns, visibleFrozenCount } = column;
       const columns = [...visibleColumns.L, ...visibleColumns.R];
-      const widths = calculateWidths(columns, this.contentsWidth);
+      const widths = calculateWidths(columns, dimension.contentsWidth);
 
       return {
         L: widths.slice(0, visibleFrozenCount),
         R: widths.slice(visibleFrozenCount)
       };
     },
+
     get offsets(this: ColumnCoords) {
       return {
         L: calculateOffests(this.widths.L, dimension.cellBorderWidth),
         R: calculateOffests(this.widths.R, dimension.cellBorderWidth)
       };
     },
+
     get areaWidth(this: ColumnCoords) {
       const { visibleFrozenCount } = column;
       const leftBorderWidth = visibleFrozenCount * dimension.cellBorderWidth;
