+++ src/view/bodyArea.tsx
@@ -2,7 +2,7 @@ import { h, Component } from 'preact';
 import { BodyRows } from './bodyRows';
 import { ColGroup } from './colGroup';
 import { Side, ColumnInfo } from '../store/types';
-import { cls } from '../helper/common';
+import { cls } from '../helper/dom';
 import { DispatchProps } from '../dispatch/create';
 import { connect } from './hoc';
 import { FocusLayer } from './focusLayer';
