+++ src/view/editingLayer.tsx
@@ -1,5 +1,5 @@
 import { h, Component } from 'preact';
-import { cls } from '../helper/common';
+import { cls } from '../helper/dom';
 
 export class EditingLayer extends Component {
   render() {
