+++ src/view/rightSide.tsx
@@ -1,5 +1,5 @@
 import { Component, h } from 'preact';
-import { cls } from '../helper/common';
+import { cls } from '../helper/dom';
 import { BodyArea } from './bodyArea';
 import { HeadArea } from './headArea';
 import { connect } from '../view/hoc';
